Cleaned up the code
Moved code to new dedicated classes, added javadocs and fixed a bug with
loading door entities.
