+++ 10x12_test.crf
@@ -1,7 +1,17 @@
 # Test level with an entity and light
 # Comments may only be present at the top and must start with a '#'
-3 2 1 1
-FLOOR WALL CORRIDOR
-CORRIDOR FLOOR FLOOR
-0 1 Key
+10 12 1 1
+WALL WALL WALL WALL WALL WALL WALL WALL WALL WALL
+WALL FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR WALL
+WALL FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR WALL
+WALL FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR WALL
+WALL FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR WALL
+WALL FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR WALL
+WALL FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR WALL
+WALL FLOOR FLOOR WALL FLOOR FLOOR FLOOR FLOOR FLOOR WALL
+WALL FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR WALL
+WALL FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR WALL
+WALL FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR FLOOR WALL
+WALL WALL WALL WALL WALL WALL WALL WALL WALL WALL
+5 6 5 Door
 25 100 25 PointLight .5/.5/.5/.5 500
